Added new tech page, reorganized the build script.
Build script now generates a page for each twig template in the root pages folder, instead of using a hard-coded array.

+++ new-site/build.php
@@ -13,6 +13,11 @@ if (!file_exists($distPath)) {
   die('Unable to write to dist folder.'.PHP_EOL);
 }
 
+$pagesPath = './pages';
+if (!file_exists($pagesPath) || !is_dir($pagesPath)) {
+  die('Missing pages folder.'.PHP_EOL);
+}
+
 // remove any files in the dist folder
 exec(sprintf(
   'rm -rf %s/*',
@@ -29,17 +34,27 @@ exec(sprintf(
 $loader = new Twig_Loader_Filesystem(__DIR__.'/pages');
 $twig = new Twig_Environment($loader);
 
-$pages = [
-  'index.html' => 'index.twig'
-];
-
-foreach ($pages as $output => $template) {
-  $outputPath = $distPath.'/'.$output;
-  $generated = $twig->render($template);
-  $minified = HTMLMinify::minify($generated);
-  if ($generated) {
-    file_put_contents($outputPath, $minified);
+if ($handler = opendir($pagesPath)) {
+  while (false !== ($pageFile = readdir($handler))) {
+    if ($pageFile === '.' || $pageFile === '..') {
+      continue;
+    } else if (is_dir($pagesPath.'/'.$pageFile)) {
+      continue;
+    }
+    $output = null;
+    if (preg_match('/^(\S+)\.twig$/', $pageFile, $output)) {
+      $outputPath = $distPath.'/'.$output[1].'.html';
+      $generated = $twig->render($pageFile);
+      $minified = HTMLMinify::minify($generated);
+      if ($generated) {
+        file_put_contents($outputPath, $minified);
+      }
+    }
   }
+
+  closedir($handler);
+} else {
+  die('Unable to read pages folder.'.PHP_EOL);
 }
 
 echo 'Done'.PHP_EOL;
